Conflicts:
	src/main/java/com/spring/controller/DcpAdminController.java
	target/classes/com/spring/controller/DcpAdminController.class
	target/classes/com/spring/dao/DcpUserDao.class
	target/m2e-wtp/web-resources/META-INF/maven/com.spring/test/pom.properties

+++ src/main/java/com/spring/util/FileDownloadView.java
@@ -44,7 +44,10 @@ public class FileDownloadView extends AbstractView{
                     e.printStackTrace();
                 }
             }
-            file.delete();
+            if((boolean) model.get("isFolder")) {
+            	file.delete();
+            }
+            
         }
         out.flush();
     }
